Update scopes
Added the scopes to the guest user and the refresh token.
This also fixes the bug created when a previousely signed in user with old scopes will not be signed out
and the refresh token will be refreshing the token for old scopes that might affect the content displayed
and the user experience.